pyproject.toml
[project]
name = "di-python-data-analyse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Eystein Waade", email = "eystein.waade@nav.no" }
]
requires-python = ">=3.12"
dependencies = [
    "db-dtypes>=1.4.3",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-bigquery-storage>=2.32.0",
    "pandas>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
